requirements.txt
easydict==1.7
graphviz==0.8.4
gym==0.10.5
imageio==2.3.0
matplotlib==2.2.2
numpy==1.14.5
Pillow==5.2.0
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
tensorboardX==2.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.23.4
torchsummary==1.5.1
torchviz==0.0.1
torchmeta==1.4.3

